Change Y values int to long

+++ app/src/main/java/com/kitzapp/telegram_stats/domain/model/chart/impl/Line.java
@@ -5,13 +5,13 @@ public class Line implements LineI {
     public static final byte TYPE_X_AXIS = 2;
 
     private String key;
-    private int[] data;
+    private long[] data;
     private byte type;
     private String name;
     private int color;
     private boolean isActive = true;
 
-    public Line(String key, int[] data, byte type, String name, int color) {
+    public Line(String key, long[] data, byte type, String name, int color) {
         this.key = key;
         this.data = data;
         this.type = type;
@@ -19,7 +19,7 @@ public class Line implements LineI {
         this.color = color;
     }
 
-    public int[] getData() {
+    public long[] getData() {
         return data;
     }
